+++ resources/views/admin/works/edit.blade.php
@@ -24,6 +24,10 @@
 
     </style>
     <link rel="stylesheet"  href="https://cdnjs.cloudflare.com/ajax/libs/dropzone/5.1.1/min/dropzone.min.css">
+    <!-- Date & Time Picker CSS -->
+    <link rel="stylesheet" href="/css/datepicker.css" type="text/css" />
+    <link rel="stylesheet" href="/css/timepicker.css" type="text/css" />
+    <link rel="stylesheet" href="/css/daterangepicker.css" type="text/css" />
 
 @stop
 
@@ -67,19 +71,39 @@
                 {!! Form::select('workstag_list[]',$workstags,$workstags_list,['class'=>'form-control js-example-basic-multiple js-states','multiple'=>'multiple']) !!}
             </div>
             <div class="form-group">
+                {!! Form::label('date','Date') !!}
+                {!! Form::text('date',null,['class'=>'form-control tleft mnth']) !!}
+            </div>
+            <div class="form-group">
                 {!! Form::label('star','Stars') !!}
                 {!! Form::select('star',['1' => '1','2'=>'2','3'=>'3','4'=>'4','5'=>'5'],null,['class'=>'form-control  js-states','id'=>'stars']) !!}
             </div>
-
+            <div class="form-group">
+                {!! Form::label('url', 'URL:') !!}
+                {!! Form::text('url', null, ['class'=>'form-control'])!!}
+            </div>
 
 
         </div>
 
 
         <div class="col-sm-6">
+
+
             <div class="form-group">
-                {!! Form::label('url', 'URL:') !!}
-                {!! Form::text('url', null, ['class'=>'form-control'])!!}
+                {!! Form::label('thumb', 'Thumbnail:') !!}
+                <div class="input-group" >
+               <span class="input-group-btn">
+                 <a id="lfm1" data-input="thumbnail" data-preview="holder1" class="btn btn-primary">
+                   <i class="fa fa-picture-o"></i> Choose
+                 </a>
+               </span>
+                    {!! Form::text('thumbnail', null,['class'=>'form-control ','id'=>'thumbnail'])!!}
+
+                </div>
+                <img id="holder1" style="margin-top:15px;max-height:100px;" src="{{$work->thumbnail}}">
+
+
             </div>
 
 
@@ -88,11 +112,11 @@
                 {!! Form::label('banner', 'Banner:') !!}
                 <div class="input-group" >
                <span class="input-group-btn">
-                 <a id="lfm" data-input="thumbnail" data-preview="holder" class="btn btn-primary">
+                 <a id="lfm" data-input="banner" data-preview="holder" class="btn btn-primary">
                    <i class="fa fa-picture-o"></i> Choose
                  </a>
                </span>
-                    {!! Form::text('banner', null,['class'=>'form-control ','id'=>'thumbnail'])!!}
+                    {!! Form::text('banner', null,['class'=>'form-control ','id'=>'banner'])!!}
                     {{--<input id="thumbnail" class="form-control" type="text" name="img">--}}
                 </div>
                 <img id="holder" style="margin-top:15px;max-height:100px;" src="{{$work->banner}}">
@@ -167,19 +191,26 @@
 
         <div class="form-inline">
 
+            @if (Route::has('login'))
+                @if (Auth::user()->isAdmin())
+                    <div class="form-group">
+                        {!! Form::submit('Update Posts', ['class'=>'btn btn-primary']) !!}
+                        {!! Form::close() !!}
+                    </div>
 
-            <div class="form-group">
-                {!! Form::submit('Update Posts', ['class'=>'btn btn-primary']) !!}
-                {!! Form::close() !!}
-            </div>
 
 
+                    <div class="form-group">
+                        {!! Form::open(['method'=>'DELETE','action'=>['AdminWorksController@destroy',$work->id],'class'=>'delete_confirm']) !!}
+                        {!! Form::submit('Delete', ['class'=>'btn btn-danger']) !!}
+                        {!! Form::close() !!}
+                    </div>
+                @else
+                    <div class="btn btn-danger">Guest cannot pass!</div>
+                @endif
+            @endif
 
-            <div class="form-group">
-                {!! Form::open(['method'=>'DELETE','action'=>['AdminWorksController@destroy',$work->id],'class'=>'delete_confirm']) !!}
-                {!! Form::submit('Delete', ['class'=>'btn btn-danger']) !!}
-                {!! Form::close() !!}
-            </div>
+
 
         </div>
 
@@ -224,6 +255,22 @@
     {{--});--}}
     {{--</script>--}}
     <script src="/js/jquery.barrating.min.js" type="text/javascript"></script>
+    <!-- Date & Time Picker JS -->
+    <script type="text/javascript" src="/js/moment.js"></script>
+    <script type="text/javascript" src="/js/datepicker.js"></script>
+    <script type="text/javascript" src="/js/timepicker.js"></script>
+
+    <!-- Include Date Range Picker -->
+    <script type="text/javascript" src="/js/daterangepicker.js"></script>
+    <script>
+        $(function(){
+            $('.mnth').datepicker({
+                autoclose: true,
+                minViewMode: 1,
+                format: "mm/yyyy"
+            });
+        });
+    </script>
 
     <script type="text/javascript">
         $(function() {
@@ -248,7 +295,9 @@
     <script>
         var domain = "";
         $('#lfm').filemanager('image', {prefix: domain});
+        $('#lfm1').filemanager('image', {prefix: domain});
     </script>
+
     <script type="text/javascript">
