+++ resources/views/admin/workstags/edit.blade.php
@@ -43,20 +43,26 @@
                         {!! Form::label('name','Name:') !!}
                         {!! Form::text('name',null, ['class'=>'form-control']) !!}
                     </div>
+                    @if (Route::has('login'))
+                        @if (Auth::user()->isAdmin())
+                            {!! Form::submit('Update Tag', ['class'=>'btn btn-primary col-sm-6']) !!}
 
 
-                    {!! Form::submit('Update Tag', ['class'=>'btn btn-primary col-sm-6']) !!}
+                            {!! Form::close() !!}
 
 
-                    {!! Form::close() !!}
+                            {!! Form::open(['method'=>'DELETE','action'=>['AdminWorkstagsController@destroy',$workstag->id]]) !!}
 
 
-                    {!! Form::open(['method'=>'DELETE','action'=>['AdminWorkstagsController@destroy',$workstag->id]]) !!}
+                            {!! Form::submit('Delete Category', ['class'=>'btn btn-danger col-sm-6']) !!}
+
+                            {!! Form::close() !!}
+                        @else
+                            <div class="btn btn-danger">Guest cannot pass!</div>
+                        @endif
+                    @endif
 
 
-                    {!! Form::submit('Delete Category', ['class'=>'btn btn-danger col-sm-6']) !!}
-
-                    {!! Form::close() !!}
 
                 </div>
             </div>
